Cleaning and extracting date helpers
