+++ src/charts/tooltip.js
@@ -7,10 +7,7 @@ define(function(require){
     const d3Transition = require('d3-transition');
     const d3TimeFormat = require('d3-time-format');
 
-    const {
-        axisTimeCombinations
-    } = require('./helpers/constants');
-
+    const {axisTimeCombinations} = require('./helpers/constants');
     const {
         formatIntegerValue,
         formatDecimalValue,
